refactor(collector): refactor css at collector item layout (#1311)
* refactor: temp update collector item design


* refactor: refactor css at collector item


* fix: fixed plugin css layout


---------
